mnt/schema reorg (#831)
* ♻️ move schemas

* 👨‍💻 move schemas

* 🧹lint
