+++ .changeset/rich-canyons-poke.md
@@ -1,0 +1,7 @@
+---
+'@curvenote/scms-server': patch
+'@curvenote/scms-core': patch
+'@curvenote/scms': patch
+---
+
+Moving schemas to central location
