+++ packages/scms-server/src/backend/uploads/remove.server.ts
@@ -2,7 +2,7 @@ import { data as dataResponse } from 'react-router';
 import { safeWorkVersionJsonUpdate } from '../occ.server.js';
 import { TrackEvent, coerceToObject } from '@curvenote/scms-core';
 import type { FileMetadataSection } from '@curvenote/scms-core';
-import type { WorkVersionMetadata } from '../metadata.js';
+import type { WorkVersionMetadata } from '../../schemas/work-version/index.js';
 import { StorageBackend } from '../storage/backend.server.js';
 import { File } from '../storage/file.server.js';
 import { KnownBuckets } from '../storage/constants.server.js';
@@ -122,7 +122,7 @@ export async function workVersionUploadRemove(
         };
 
         // Remove the file from metadata
-        delete updatedMetadata.files[path];
+        delete updatedMetadata.files?.[path];
 
         return updatedMetadata as any;
       },
