+++ packages/scms-server/src/schemas/index.ts
@@ -1,0 +1,2 @@
+export * from './files.js';
+export * from './work-version/index.js';
