+++ platform/scms/package.template.json
@@ -117,6 +117,7 @@
     "@app-config/config": "^2.8.7",
     "@app-config/main": "^2.8.7",
     "@app-config/vite": "^2.8.7",
+    "@curvenote/blocks": "^1.6.2",
     "@curvenote/cdn": "^0.4.0",
     "@curvenote/check-definitions": "^0.1.2",
     "@curvenote/check-ui": "^0.6.9",
